+++ src/Symfony/Component/Validator/Constraints/Locale.php
@@ -30,14 +30,10 @@ class Locale extends Constraint
     ];
 
     public $message = 'This value is not a valid locale.';
-    public $canonicalize = false;
+    public $canonicalize = true;
 
     public function __construct($options = null)
     {
-        if (!($options['canonicalize'] ?? false)) {
-            @trigger_error('The "canonicalize" option with value "false" is deprecated since Symfony 4.1, set it to "true" instead.', E_USER_DEPRECATED);
-        }
-
         if (!class_exists(Locales::class)) {
             throw new LogicException('The Intl component is required to use the Locale constraint. Try running "composer require symfony/intl".');
         }